[IMPL] new associator + now we add chains of sides
Two key ideas of new associator:

1. We must project edge to the shade contour
and calculate areas of parts on which it divides
the contour. If one of them is small, we add
associations

2. Add not one side, but multiple sides, since
it is possible that the edge is projected on
multiple sides. We must take into account small
sides and adjacent sides (not to add them).